Softland - HTML Responsive Landing Template
Softland is a fully responsive, clean and modern high-converting landing page template. It is a beautifully handcrafted, pixel perfect HTML5 landing page template based on the latest twitter bootstrap v4.5. Softland comes with 6 home demos Html files. All files are organized we believe it will be easy to use and edit them. This template is well organized and very easy to customize. Quick and easy to setup.

https://www.codester.com/items/24626/softland-html-responsive-landing-template